Modularize skills view and desktop backend
